Move ScannedFeature and Literal* to separate files
