+++ src/ast/ast.ts
@@ -32,12 +32,13 @@
  * Earlier stages have the longer names, like ParsedDocument and ScannedElement.
  */
 
-export * from './descriptor';
 export * from './document-descriptor';
 export * from './element-descriptor';
 export * from './event-descriptor';
 export * from './import-descriptor';
 export * from './inline-document-descriptor';
+export * from './literal';
 export * from './property-descriptor';
 export * from './resolvable';
+export * from './scanned-feature';
 export * from './source-range';
